ResourceManager: Rename 'entry_slist' param in load_context function.
The name of this variable doesn't tell much about what it's holding.
It's definitely a GSList, holding 'entries' but since we separate
transient objects and sessions into separate data structures what we
want to know is what sort of object this list holds. Since it's used to
hold transient objects we rename it to 'loaded_transients'.

Signed-off-by: Philip Tricca <flihp@twobit.org>