ResourceManager: Rename structures holding transient objects for clarity
The is an artifact of the initial implementation that only managed
transient objects. There was no need to differentiate between transient
objects and sessions back then and so the name is a bit too generic.
Renaming this function processing the transient objects and the
structure that we use to store them serves to make the code more
reasable.

Signed-off-by: Philip Tricca <flihp@twobit.org>