SessionList: Remove _lookup_connection function.
This function looks identical to the _lookup_handle function but it
behaves very differently. This inconsistency is annoying and the
function itself is awkward. The same functionality can be achieved by
using the current foreach iterator. This produces an extra layer of
indirection when processing closed connections but it's a small price to
pay to get rid of a bad function.

Signed-off-by: Philip Tricca <flihp@twobit.org>