Tpm2Response: Add check for response code to _has_handle logic.
This missing check came from an effort to clean up the ResourceManager
code that identifies handles newly allocated by the TPM2 as they're
returned to the caller. This code had a check for an RC indicating
success as well as a call to tpm2_response_hash_handle. The response
can't have a handle however unless the associated command succeeded.

The most natural thing to do in this case was to just add this check to
the tpm2_resposne_has_handle. Inverting the existing logic in the
function allows us to make all of the required checks in a single
conditional.

This commit also rewrites the comment documentation for this function to
describe the new logic.

Signed-off-by: Philip Tricca <flihp@twobit.org>