ResourceManager: Fix state logic when loading sessions from handle area.
This bad logic was a hold over from some old and outdated code
implementing an early prototype for the "extended session" code. The
logic that loads the sessions referenced in a command must always be in
the SESSION_ENTRY_SAVED_RM state. The logic removed was an attempt to
allow a client to automatically take over a saved session from another
connection which was a terrible idea that was never fully implemented.

Signed-off-by: Philip Tricca <flihp@twobit.org>