ResourceManager: Replace list tracking loaded sessions with new state.
Previously the ResourceManager tracked which sessions were loaded by
storing the relevant SessionEntry in a GList. After a command was
completed each SessionEntry in this list was removed and each session
cleaned up. This kept us from having to iterate through the whole list
of SessionEntries (an optimization) but it made some of the logic in the
ResourceManager awkward. This was a premature optimization.

The code is much more simple if we add a new state to the
SessionEntryStateEnum for a session when it's LOADED. After each command
is done we can then interate over thw list of all sessions cleanup up
after all of the loaded sessions. We lose the optimization from the
previous approach but the code is a lot easier to read and understand.

Signed-off-by: Philip Tricca <flihp@twobit.org>