ResourceManager: Fix race condition in session save logic.
The post processing function that saves all loaded sessions doesn't lock
the SessionList before iterating over it. In cases where a command has
loaded sessions if the connection is closed immediately after the
command completes it's likely that the connection teardown logic will
end up flushing a session that the post-processing function is in the
processing of saving.

The solution to this is to simply lock the SessionList before the
post-processing function starts saving sessions.

Signed-off-by: Philip Tricca <flihp@twobit.org>