tcti-dynamic: Make the tss2 device TCTI library SONAME the default
The raw libtss2-tcti-device.so is only included in the development package
but the tpm2-abrmd tries to dynamically load this instead of the library's
SONAME which is included in the tpm2-tss package.

This was reported in [0] and partially fixed by the commit d2f0b2d1f0f
("tcti-util: Use proper SONAME for TCTI dlopen"). But I still needed these
changes so the tpm2-abrmd daemon could lookup the correct library name.

[0]: https://lists.01.org/pipermail/tpm2/2018-May/000700.html

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>