ci: Reorganization of compiler specific steps.
Our travis.yml runs a number of unnecessary build steps that are slowing
down the CI builds unnecessarily:
- executing the check-code-coverage target is unnecessary, the coveralls
tool will do this for us
- when running scan-build we run the distcheck target twice

This patch better separates the two compiler specific build steps:
1) scan-build with clang
2) code-coverage megrics collected with gcc

We have to do some scripting to determine the results of scan-build and
the results from the test harness. We do so by parsing the
test-suite.log and fail if we don't match a regex.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>