ResourceManager: Fix race between session save and connection termination.
0115d572bc4f6026a28a2ed5d7340ca77c6dc4d2 fixed the worst part of this
but a very narrow window for a race persisted. For it to trigger the
client must terminate their connection to the daemon triggering a call
to the resource_manager_on_connection_removed function by an external
thread before the main processing thread saves all of the sessions
associated with the previous command.

Fixing this requires that the main processing loop hold the lock on the
SessionList until all sessions have been saved.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>