CommandSource: Replace 'connection-removed' signal with ControlMessage.
When a Connection is removed we need a mechanism to inform the
ResourceManager so that it can cleanup associated resources. Previously
we used the 'connection-removed' event produced by the
ConnectionManager. This complicates the ResourceManager requiring
synchronization across the internal thread and the thread invoking the
callback / signal.

This commit removes our use of this signal and replaces it with a new
ControlCode within the ControlMessage object: CONNECTION_REMOVED. When
a Connection is torn down the CommandSource creates a new ControlMessage
with this 'code' and a reference to the relevant Connection object. The
downstream sinks do whatever processing they need and then pass the
message along.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>