SessionList: Remove all locking.
I got very little "right" with this locking mechanism. It's no longer
required and so this commit removes it completely. This was the wrong
granularity for locking what is effectively state within a higher-level
object (ResourceManager) and was required for a poorly designed
connection tear-down mechanism that has since been fixed.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>