ci: Build 'check' target when CC is gcc.
This is to collect code coverage metrics. The distcheck target will
clean up after itself and delete these. We still execute distcheck for
clang builds, now for gcc we build the 'check' target so that the test
coverage metrics can be sent to coveralls and we don't incur the cost
of two builds.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>