build: Don't run integration tests If --enable-integration is omitted.
This adds the tests in TESTS_INTEGRATION to the TESTS variable only when
'ENABLE_INTEGRATION' is defined. Additionally the integration test
utility library libtest is added to noinst_LTLIBRARIES now in this
conditional as well. This keeps us from wasting cycles building the
library when the integration tests are disabled.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>