build: Cleanup logic enabling test cases based on configuration options.
The configure script will set ENABLE_INTEGRATION if the integration
tests should be run (--enable-integration). HWTPM is set if the
integration tests are to be run against the hardware TPM instead of the
simulator. These two conditionals are a bit tangled in that we only care
about HWTPM if ENABLE_INTEGRATION has been set as HWTPM will only be set
if ENABLE_INTEGRATION has been set as well.

This patch moves the HWTPM checks so that they are only in the scope of
the ENABLE_INTEGRATION block.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>