test: Add test to trigger context gap.
To do this we must cause the session context sequence value for saved
session to exceed the TPM_PT_CONTEXT_GAP_MAX value. This requires that
we create and save a context to establish the initial sequence value and
then create, save and flush sessions until the difference between the
initial and current sequence number exceeds TPM_PT_CONTEXT_GAP_MAX. If
we're able to do this without getting the TPM_RC_CONTEXT_GAP RC from the
TPM then the RM is handling the gap for us and the test passes.

If this test is run against the RM it should pass. If run agains the
simulator direclty it should fail.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>