Fix gdbus-codegen lookup for recent versions of GLib
If GLib is built using Meson, the gdbus_codegen variable contains the
whole path to the binary instead of just the binary name. This cannot be
parsed by AC_PATH_PROG and leads to an error in the configure script.
According to https://gitlab.gnome.org/GNOME/glib/issues/1521#note_313402,
the recommended way to solve this is to look up gdbus-codegen directly
without using the pkg-config variable.