ResourceManager: Move abandoned session tracking to the SessionList.
Having this logic in the RM was pretty awkward. Additionally it required
that a lot of the code for managing the state of SessionEntry objects
live in the RM when it's a better fit for an object dedicated to
managing them (the SessionList).

This commit refactors this such that the queue tracking abandoned
sessions lives in the SessionList. This allows us to move most of the
code around accounting for SessionEntry object state can move there as
well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>