travis.yml: Existence of build directory indicates build failure.
scan-build makes this a bit complex but it's not too bad. This fixes a
bug in the logic used previously that fails the build if the test log
file isn't available. This naively assumption doesn't account for the
fact that a successful build will have removed this file by the time we
check. Instead we should fail if the build directory exists. The build
won't remove the build directory if the build fails so we can debug
stuff.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>