build: Tighten up the sanity test for the GDBUS_CODEGEN variable.
If the configure script is invoked with this variable set by the caller
to something nonsensical (not an executable etc) the sanity test (which
only tests for -z) will pass but the build will fail. We can tighten
this check up to ensure that the value of this variable is set to an
executable. For this check to work properly we must lookup the full path
to the executable (due to c0b0b70).

We could take this a step further by ensuring the executable is able to
transform some known input in the same way autoconf checks that the
compiler / CC "works". This seems like overkill currently.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>