ci: Disable documentation in tpm2-tss build.
The upstream tpm2-tss build recently made doxygen a hard requirement (a
very good thing). This causes breakage for us since we weren't
installing doxygen in the CI image and we hadn't explicitly disabled it
either. This commit disables the documentation build for this dependency
for the sake of efficiency.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>