AccessBroker: If TCTI receive fails free memory allocated for command.
If a function fails and it has previously allocated memory it should free
it. This function is only called from one place and this calling
function was detecting the failure and freeing the memory. This is very
awkward and will confuse people like it just did to me without this fix.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>