ResourceManager: Dial up logging when saving contexts fails & check TPM RC.
When the RM is cleaning up after a command (response already sent to the
client at this point) by ContextSaving loaded sessions the command may
fail. TPM2_RC_CONTEXT_GAP is a good example but the TCTI may fail as
well. Log TCTI failures as 'ctitical' and command failures as 'warning'.

Additionally by checking the RC from the TPM we fix a bug where an
attempt to use a session after the RM failed to save it resulted in a
nonsensical response.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>