ResourceManager: Provide typed interface to save session context function
The current function is only useful as a callback from a GLib-style
foreach function. Calling this interface directly from the code (instead
of as a callback) hurts readability. This function refactors the
function into a proper C function and a wrapper that is useful as a
callback from a foreach function.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>