tcti: Refactor TCTI classes to use a factory pattern.
This commit cleans up some of the oldest code in the project to more
effectively use the dynamic TCTI initialization mechanisms. We've
supported the dynamic init API / protocol for a while now but I put off
going back to pay down this technical debt given the scope of the work.
Can't put it off any longer though.

The Tcti class is now a concrete class instead of an abstract one. The
use of an abstract class was necessary when the TCTI libraries didn't
support the dynamic initialization mechanism. When this was first added,
all concrete implementations of this class were replaced with a single
TctiDynamic class since the only unique code was the initialization
calls. Using this single class derived from the abstract base was just
my way of avoding this cleanup.

A new class called TctiFactory now encapsulates the Tcti creation /
initialization logic. When it's constructed we pass it the TCTI name and
configuration string (all that's necessary to instantiate a TCTI
context). We then call the 'tcti_factory_create' function to create
instances of the Tcti class that wrap a TSS2_TCTI_CONTEXT.

The rest of the changes in this commit are associated changes in the
test harness. These changes replace the old EchoTcti / libtss2-tcti-echo
with a much more simple 'mock' tcti that can be used directly with the
concrete Tcti & TctiFactory.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>