build: Stop including libtest in LIBADD for tcti lib.
The automake noinst_LTLIBRARIES variable was being used to include
the utility library for linking with the TCTI library. When the
integration tests are enabled they must be  added to this variable as
well which causes it to be linked into the TCTI library unnecessarily.
This bloats the shared object. This patch removes the problematic use of
noinst_LTLIBRARIES using libutil directly instead.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>