ci: Update travis dist from 'trusty' container to 'xenial' VM.
travis-ci announced their plans to migrate all users to a single
infrastructure. This boils down to them dropping support for the
container infrastructure. Details are here:
https://blog.travis-ci.com/2018-10-04-combining-linux-infrastructures

The container infrastructure was great since the build queues seem to
clear much more quickly but if it's going away we need to migrate. The
good news is that the xenial image is a much needed upgrade. This new
image allows us to drop the cmocka source build in favor of the xenail
package.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>