Mask object addresses in debug output.
Dumping the address of a GObject in debug output is useful as a
mechanism for tracking objects. This effectively uses the address of the
object as an ID value. This is however undesirable as there is a
possibility that a system may be deployed with debug output turned on
mistakenly. This would case the log file to contain info that may be
used to weaken ASLR.

To combat this we add a new function to the util module that we use to
mask object addresses. This is done by combining the address with a
random value using the bitwise exclusive or (aka XOR) operation. The
random value / seed is initialized by adding another new function to
initialize the 'util' module. This function must be called before any
addresses are masked.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>