build: Install aclocal macros to $(srcdir)/m4.
When autoreconf is invoked the required m4 files are included in the
build by making symlinks under $(srcdir)/m4 to the installed versions.
When building distribution tarballs these files are copied and included.
This commit adds the `--install` option when invoking autoreconf. This
causes these macro source files to be copied into $(srcdir)/m4 (instead
of symlinks) which causes a git source tree / build to be more
consistent with the distributed tarballs.

Signed-off-by: Alon Bar-Lev <alon.barlev@gmail.com>
Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>