build: update for ax_code_coverage.m4 version 2019.01.06
@CODE_COVERAGE_RULES@ doesn't exist any more and needs to be replaced.
Also includes a compatibility wrapper for older versions of the file.

There is a bug in the new version of ax_code_coverage.m4
(cf. https://github.com/autoconf-archive/autoconf-archive/pull/182):
AM_DISTCHECK_CONFIGURE_FLAGS is recursively defined, which causes
"make distcheck" to fail in combination with --enable-code-coverage.
This is fixed by overwriting the faulty definition.

Signed-off-by: Jonas Witschel <diabonas@gmx.de>