ci: Fix coveralls.
Coveralls integration was broken by some changes around
54035b7ea2d58c8dd2337c032b0d7bc5b4ee4a62. This commit restores coveralls
to a functional state. Code coverage is only enabled for builds with
clang. The clang build has also been changed to build the 'check' target
instead of 'distcheck'. This is required because a successful
'distcheck' build will cause the build directory to be deleted before we
can run coveralls (the source of the original issue).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>