test: Mock remaining dbus interaction out of TCTI unit test.
The test was missing mock function for the code in the gdbus-codegen
generated module. This is the initialization function that creates the
connection with dbus: tcti_tabrmd_proxy_new_for_bus_sync. This commit
adds the necessary mock function and sets up the mock stack for the
relevant functions.

The changes to the test setup function are necessary because we must
pass a valid GDBusProxy object to the mock function. This requirement
comes from the fact that a successfull call to this function will
trigger code that casts this pointer using the G_DBUS_PROXY macro.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>