tcti-factory: Add check for null after TCTI initialization.
The code checks the return value and assumes that the context has been
set to a non-NULL value if the response code indicates success. Most
static analysis tools don't look down into the 'tcti_util_dynamic_init'
function to see that this is safe however. Adding an explicit check
comes at nearly no cost and makes the code clear. Some cleanup was
required in the unit test that relied on the old behavior as well.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>