test: remove start-auth-session.int.c from the XFAIL list
The test creates a single POLICY session and then calls Tss2_Sys_GetCapability()
to report on the number of session handles.  The test failure was caused by
using TPM2_ACTIVE_SESSION_FIRST instead of TPM2_LOADED_SESSION_FIRST and
vice-versa.  LOADED_SESSION_FIRST is defined as HMAC_SESSION_FIRST and
ACTIVE_SESSION_FIRST is defined as POLICY_SESSION_FIRST.  During the test we
should expect 0 of the former and 1 of the latter.

Signed-off-by: Jeffrey Ferreira <jeffpferreira@gmail.com>