tabrmd-options: Fix bug introduced in c5e66b78de.
The comment above this function clearly states the reason for these
return values. This was however missed in this refactoring. This fix
includes a small change to free the GOptionContext as early as possible.
This allows us to eliminate the final conditional returning the status
from the last function call directly.

This allows us to eliminate one of the unit tests without affecting the
coverage metrics.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>