tcti: Change function visibility to the default & call directly in units.
This will not cause these functions to be be exposed by the shared
object. The linker version script (aka map file) controls this. Doing
this does however enable testing of the functions by unit tests. This
required minor fixups to the unit tests due to disagreement about the
appropriate return code when the context paramter is NULL for a few
functions.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>