tcti: fix partial reads
This fixes a bug in the TCTI that causes a hang when the caller
implements the pattern referred to as 'partial reads' in tpm2-tss. The
issue was caused by the implementation not properly handling an edge
case where the response buffer is only the 10 byte TPM response header.
The TCTIs use of the read_data function from the util module was also
replace with a new function that does not retry after short reads. This
pattern works well for the daemon since they're used by a dedicated
thread intended to block.

This commit includes the necessary cleanup in the associated unit test
and some refactoring to move the relevant tests (all related to the
'receive' function) to a new unit test. This is the first step in
decomposing the oversized TCTI unit test (> 1kLOC).

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>