build: Simplify unit test CFLAGS and LIBADDs.
This combines all CFLAGS and LIBADDs for the unit tests into a variable
for each. This causes the test executables to be over linked in some
circumstances. These binaries however are not installed as part of the
build process and so this should be viewed as a choice of simplicity
over accuracy where sacrificing some accuracy is acceptable.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>