util: replace xor address obfuscation / objid with truncated sha256
xor worked fine for this purpose but the crypto folks get itchy if we
use anything other than a standard hash algorithm. Luckily glib has a
sha256 implementation that I didn't know about till now. This function
must truncate the hash to a void* so we increase the probability of a
collision. Since this value is never used for anything other than
debug output it's sufficient for the purpose.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>