remove unneeded endian.h include
The <endian.h> header can be found in different locations
on different OSes. On FreeBSD one needs to include <sys/endian.h>
while on Linux the file ot include is <endian.h>. To make this build
correctly on different OSes we can wrap the include in appropriate #ifdefs
or just remove the explicite include and the correct file will be pulled
by the <tss2/tss2_mu.h> header anyway.

Signed-off-by: Tadeusz Struk <tadeusz.struk@intel.com>