CommandSource: Use glib macro to cast to GSourceFunc.
Casting to the GSourceFunc type from custom callback functions is common
in glib. glib provides a macro to perform this cast and so we should use
it. This also quiets warnings from newer compilers about casting to an
incompatible pointer type.

For systems still shipping glib versions < 2.58 we provide an
alternative implementation locally.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>