tabrmd-init: Give meaning to return values from tabrmd-init.
This function previously returned 0 for success and 1 for failure. But
due to the systemd Restart=always policy failures were ignored and the
daemon would always be restarted if it failed. We've seen a number of
people installing tabrmd on systems without TPMs and this causes the
initialization of the TCTI to fail. This results in a restart loop which
is very undesirable.

This patch adds meaningful response codes to a number of possible
failure conditions in the daemon init function. These return codes are
eventually used as the exit status for the daemon. Finally we instruct
systemd to *not* restart the daemon if any of these return codes are
produced since none of these are recoverable errors. The new response
codes uses are as follows:

EX_OSERR - used if: adding signal handlers fail, seeding the random
number object from a file fails, a thread fails to start
EX_IOERR - used if: initialization of the TCTI fails
EX_UNAVAILABLE - used if: attempts to initialize the TPM fails

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>