test: replace netstat by ss
netstat is deprecated in favour of ss, see the netstat man page. Also
restrict listing to IPv4 to fix a rare bug when binding only succeeds
for the IPv6 interface, but not for the IPv4 interface which is used by
the tests to connect to the simulator
(https://github.com/tpm2-software/tpm2-tss/pull/1234).

Signed-off-by: Jonas Witschel <diabonas@gmx.de>