tabrmd-init: Exit with EX_IOERR when IPC fails to connect.
When the IPC mechanism fails to connect the tabrmd-init module receives
an event from the IpcFrontend object and then kills off the GMainLoop.
The current implementation provides no way for the main function to know
why the loop was killed off and so the it exits with a status of 0
(EXIT_SUCCESS) which causes systemd to restart the daemon. This causes a
restart loop.

This patch adds a boolean to the gmain_data_t structure that the
callback sets when the IPC disconnects. This allows the main function to
check for the flag when the GMainLoop terminates and return a meaningful
exit code (EX_IOERR). This exit code will prevent systemd from trying to
restart the daemon.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>