tpm2-abrmd.service: Don't restart the service if fails
The service has the option to always restart if it fails, but this is not
really needed since the service is D-Bus activated.

So if the daemon can't be restarted due an unrecoverable condition, there
is no point to attempt to do it. And if it can be started, then it will be
done when a client does a D-Bus call again.

Fixes: #673

Signed-off-by: Javier Martinez Canillas <javierm@redhat.com>