systemd: Use a specific device dependency instead of udev-settle
Pulling in udev-settle affects the entire system – it makes the entire
boot process block until all events have been processed. This is
usually overkill when we only need one specific device.

(It is not possible to depend on a wildcard /dev/tpm*, but it is a safe
assumption that systems having a real TPM will call it tpm0.)

Signed-off-by: Mantas Mikulėnas <grawity@gmail.com>