build: Integrate SELinux policy module into autotools build.
This commit adds the `--with-sepolicy` option to the configure script.
If provided (disabled by default) the SELinux policy module is built
and installed as part of the default make target. This option takes a
parameter (optional) that is the path to 'include' directory provided
with the reference policy development package. If omitted the default is
/usr/share/selinux/devel/include.

Unfortunately the Makefile provided by the reference policy that we had
used previously breaks VPATH builds. In its place we implement the
required build rules manually fixing the VPATH incompatibilities. This
new mechanism for building the policy module replaces the old Makefile
previously available under $(srcdir)/selinux. Build commands may no
longer be executed under the $(srcdir)/selinux directory. Instead the
SELinux policy module is built and installed like any other artifact.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>