build: Fix broken attempt to add -Wformat-security to CFLAGS.
The 'format-security' warning is dependent on the 'format' warning. It
cannot be used without first enabling 'format' warnings. By adding
'-Wall' to the flags used in the AX_CHECK_COMPILE_FLAG macro we enable
all warnings including 'format'.

This likely has the side effect of making other checks more strict but
IMO that's a good thing.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>