test: Refactor access-broker GetCapability mock & cover handle query.
Previously the mock function for Tss2_Sys_GetCapability was specific to
the tagged tpm properties capability. This prevented coverage of other
uses in the access-broker module like queries for handles. This commit
refactors this mock function to a form flexible enough to allow any
capability to be queried. For now we only support the two we need:
tagged properties and handles.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>