access-broker: Refactor system API init function.
This function was in bad shape. This commit removes an unnecessary NULL
check, adds asserts for parameters that MUST not be NULL, and replaces a
g_error call that should be a warning.

Getting good test coverage for this function required adding a simple
mock function for Tss2_Sys_Initialize.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>