CommandAttrs: Move interaction with TPM into AccessBroker.
The AccessBroker has become our centralized location for use of
tss2-sys. This commit moves the call to GetCapability from the init
function in CommandAttrs to a specialized function in the AccessBroker.
This new function returns a newly allocated array of TPMA_CC objects.
The length of this array is returned through the 'count' variable. The
caller is responsible for deallocating the TPMA_CC array when it's no
longer in use.

Addtionally we fixup the CommandAttrs and AccessBroker unit test and
remove a now unused function from the AccessBroker.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>