AccessBroker: Rename to Tpm2.
The AccessBroker was misnamed long ago. This object is our interface to
the TPM2 device. It takes ownership of the Tcti object when it's created
and it simplifies calling various TPM commands when the daemon is
started as well as transmitting and receiving Tpm2Command & Tpm2Response
objects.

Future attemps to build access control mechanism (something that an
AccessBroker would be responsible for) should likely be done as a filter
in the command processing pipeline.

Signed-off-by: Philip Tricca <philip.b.tricca@intel.com>