Makefile: add glib and asan options to ENV
When debugging with glib, it helps to have the ASAN_OPTIONS set to get
the full stack trace. It's also imperitive that we disable the slice
allocator becuase the library won't free the memory on exit. Thus, malloc's
may not always have a free and will trigger a memory leak.

Signed-off-by: William Roberts <william.c.roberts@intel.com>