tabrmd-init.c: fix leaks on main to thread tpm2 instance
Theirs a case where the Tpm2 object coming in from main to the thread
fails setup and the cleanup function doesn't unref it. Move it to the
main cleanup routine and use g_clear_object to be *clear* on whom owns
the reference.

Fixes leaks like:
Indirect leak of 4176 byte(s) in 1 object(s) allocated from:
    #0 0x7f652e71cdc6 in calloc (/usr/lib/x86_64-linux-gnu/libasan.so.5+0x10ddc6)
    #1 0x7f652e25ad30 in g_malloc0 (/usr/lib/x86_64-linux-gnu/libglib-2.0.so.0+0x57d30)
    #2 0x555ebb1a1c5f in sapi_context_init src/tpm2.c:162
    #3 0x555ebb1a2fa8 in tpm2_new src/tpm2.c:438
    #4 0x555ebb19d665 in init_thread_func src/tabrmd-init.c:178
    #5 0x555ebb19bede in init_thread_func_tpm2_init_fail test/tabrmd-init_unit.c:199
    #6 0x7f652e6074e0  (/usr/lib/x86_64-linux-gnu/libcmocka.so.0+0x54e0)

Signed-off-by: William Roberts <william.c.roberts@intel.com>