Initialize tcti in advance
init_thread_func creates a thread and then initializes the tcti.
When /dev/tpm0 does not exist, the thread is created. If other processes communicate with abrmd, the connection may be established. However, /dev/tpm0 does not exist, and init_thread_func releases resources. After the resources are released randomly, the resources may still be used by id_pid_mix_from_incall to randomly obtain the kernel dump in uint64.
The patch initializes the tcti in advance to avoid concurrency.

https://github.com/tpm2-software/tpm2-abrmd/issues/838